Move S2167 test source file to compiling package
